requirements.txt
numpy>=1.21.0
pandas>=1.3.0
yfinance>=0.2.20
scikit-learn>=1.0.0
tensorflow>=2.9.0
flask>=2.0.0
# Fixed Web3 Dependencies for IndiCoin Project
# These versions are tested to work together

web3==6.11.0
py-solc-x==2.0.2
pytest==7.4.0
pytest-asyncio==0.21.1
eth-account==0.9.0
eth-utils==2.2.0
eth-typing==3.5.0
hexbytes==0.3.1

# Additional stable dependencies
requests==2.31.0
setuptools==68.0.0
